+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Data state `id` property split into 2 `rowId` and `columnId`.
 - Remove `actions`, `pipe`, `selectionKey`, `filterFetch` input properties from the grid component.
 - Remove `keep` method from grid plugin.
+- Data rowId returns row index by default not the instance.
 
 ### Added
 - Ivy support.